Payments in creating resumes and vacations

+++ TeleportBot/core/bot/resumes/create.py
@@ -2,9 +2,10 @@ from telegram import ParseMode
 from telegram.ext import ConversationHandler
 from core.resources import strings, keyboards
 from core.bot.utils import Navigation
-from core.services import categories, resumes
+from core.services import categories, resumes, settings
+from core.bot import payments
 
-TITLE, DESCRIPTION, CONTACTS, REGION, CITY, CATEGORIES = range(6)
+TITLE, DESCRIPTION, CONTACTS, REGION, CITY, CATEGORIES, TARIFFS, PROVIDER = range(8)
 
 
 def to_parent_categories(query, context):
@@ -137,7 +138,8 @@ def resume_city(update, context):
 
 
 def resume_categories(update, context):
-    language = context.user_data['user'].get('language')
+    user = context.user_data['user']
+    language = user.get('language')
     query = update.callback_query
     category_id = query.data.split(':')[1]
     if category_id == 'back':
@@ -154,16 +156,25 @@ def resume_categories(update, context):
         else:
             return to_parent_categories(query, context)
     if category_id == 'save':
-        resumes.create_resume(context.user_data['resume'])
-        success_message = strings.get_string('resumes.create.success', language)
-        help_message = strings.get_string('resumes.create.success.help', language)
-        context.bot.delete_message(chat_id=query.message.chat.id, message_id=query.message.message_id)
-        context.bot.send_message(chat_id=query.message.chat.id, text=success_message)
-        menu_keyboard = keyboards.get_keyboard('menu', language)
-        context.bot.send_message(chat_id=query.message.chat.id, text=help_message, parse_mode=ParseMode.HTML,
-                                 reply_markup=menu_keyboard)
-        Navigation.to_account(update, context)
-        return ConversationHandler.END
+        if user.get(user.get('user_role')+'_tariff'):
+            payment_settings = settings.get_settings()
+            item_cost = payment_settings.get(user.get(user.get('user_role')+'_tariff'))
+            if user.get('balance_' + user.get('user_role')) >= item_cost:
+                resume = resumes.create_resume(context.user_data['resume'])
+                context.user_data['user'] = resume.get('user')
+                success_message = strings.get_string('resumes.create.success', language)
+                help_message = strings.get_string('resumes.create.success.help', language)
+                context.bot.delete_message(chat_id=query.message.chat.id, message_id=query.message.message_id)
+                context.bot.send_message(chat_id=query.message.chat.id, text=success_message)
+                menu_keyboard = keyboards.get_keyboard('menu', language)
+                context.bot.send_message(chat_id=query.message.chat.id, text=help_message, parse_mode=ParseMode.HTML,
+                                         reply_markup=menu_keyboard)
+                Navigation.to_account(update, context, new_message=True)
+                del context.user_data['resume']
+                return ConversationHandler.END
+        empty_balance = strings.get_string('empty_balance', language)
+        query.answer(text=empty_balance, show_alert=True)
+        return payments.start(update, context)
 
     category = categories.get_category(category_id)
     children_categories = category.get('categories')
